requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
requests
beautifulsoup4
pandas
lxml
apache-airflow-providers-snowflake
snowflake-connector-python
dbt-snowflake